moved inmemory lookup into production
